Geekyshows/python/2.LLMs and Chat Models with LangChain Python/ch2/requirements.txt
langchain==0.3.21
python-decouple==3.8
langchain-openai==0.3.11
